Show last_logon_time in userinfo response
- Made last_logon_time nullable so avoid displaying erroneous timestamps
  in the userinfo response.
- last_logon_time now defaults to null in the database
- Add LAST_LOGON_TIME to token ClaimConstants

[#136373683] https://www.pivotaltracker.com/story/show/136373683

+++ server/src/test/java/org/cloudfoundry/identity/uaa/user/JdbcUaaUserDatabaseTests.java
@@ -316,12 +316,12 @@ public class JdbcUaaUserDatabaseTests extends JdbcTestBase {
         when(timeService.getCurrentTimeMillis()).thenReturn(1000L);
         db.updateLastLogonTime(JOE_ID);
         UaaUser joe = db.retrieveUserById(JOE_ID);
-        assertEquals(joe.getLastLogonTime(), 1000L);
+        assertEquals((long) joe.getLastLogonTime(), 1000L);
 
         when(timeService.getCurrentTimeMillis()).thenReturn(2000L);
         db.updateLastLogonTime(JOE_ID);
         joe = db.retrieveUserById(JOE_ID);
-        assertEquals(joe.getLastLogonTime(), 2000L);
+        assertEquals((long) joe.getLastLogonTime(), 2000L);
     }
 
     @Test(expected = UsernameNotFoundException.class)
